docker_working

+++ start_ros.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 # Source ROS setup
 source /ros_entrypoint.sh
 
@@ -13,30 +12,34 @@ if [ ! -f "/ros/catkin_ws/devel/setup.bash" ]; then
     cd src
 
     # Clone necessary repositories
-    git clone https://github.com/jmscslgroup/skynetmini.git
-    git clone https://github.com/jmscslgroup/subtractor.git
-    git clone https://github.com/jmscslgroup/odometer.git
-    git clone https://github.com/jmscslgroup/carsimplesimulink.git
-    git clone https://github.com/giorgioa02/AV-T_bag.git
+    [ ! -d "skynetmini" ] && git clone https://github.com/jmscslgroup/skynetmini.git
+    [ ! -d "subtractor" ] && git clone https://github.com/jmscslgroup/subtractor.git
+    [ ! -d "odometer" ] && git clone https://github.com/jmscslgroup/odometer.git
+    [ ! -d "carsimplesimulink" ] && git clone https://github.com/jmscslgroup/carsimplesimulink.git
+    [ ! -d "AV-T_bag" ] && git clone https://github.com/giorgioa02/AV-T_bag.git
 
     # Return to the workspace root and build
-    cd /ros/catkin_ws
+    cd ..
     catkin_make
 
     # Source the new setup file
     source devel/setup.sh 
 
-    # Copy the launch file to the main workspace directory, if needed
+    # Copy the launch file to the main workspace directory
     if [ -f "src/skynetmini/launch/skynetmini.launch" ]; then
         cp src/skynetmini/launch/skynetmini.launch .
     fi
-    #Copy the bagfile out
-    cp src/AV-T_bag/mytest.zip .
-    unzip mytest.zip .
-    mv 2023_09_28_16_19_25_2T3MWRFVXLW056972microstrain_record.bag mytest.bag
-    # Launch the ROS application
-    roslaunch skynetmini skynetmini.launch
+    
+    # Copy and unzip the bag file if it exists
+    if [ -f "src/AV-T_bag/mytest.zip" ]; then
+        cp src/AV-T_bag/mytest.zip .
+        unzip mytest.zip -d .
+        mv 2023_09_28_16_19_25_2T3MWRFVXLW056972microstrain_record.bag mytest.bag
+    fi
+    # Launch the ROS application if needed (uncomment if required)
+    # roslaunch skynetmini skynetmini.launch
 fi
 
 # Start roscore if the workspace is already built and skip building
-exec roscore
+# exec roscore
+exec bash
